IE11 fixes (#21951)
* Use the "qs" library to parse the current URL searching for the dev "offline" params.

"URLSearchParams" isn't supported in IE/Edge.

* Move the required polyfills to a different file so they load before "draft-js".

All "import" statements are parsed and run before the rest of the file,
so inlining the polyfills isn't working anymore. By just moving them to a new file
and importing it before "draft-js" the problem is fixed.
